+++ HISTORY.rst
@@ -40,6 +40,7 @@ latest (unreleased)
 * BIZ-2795: Add account-financial-report-qweb for journal reporting
 * BIZ-2796: Add account-financial-report-qweb
 * ALCYN-1801: do not send "You have been assigned to..." email to new followers on any documents
+* ALCYN-1796 Add field for product.template.nb_days_out_of_stock
 
 **Bugfixes**